test/requirements.txt
pytest==8.2.2
cocotb==1.9.1
cocotbext-uart==0.1.2
